move dataswarm_client_ops to ds_client_ops
