+++ dataswarm/manager/ds_client_ops.h
